# Deterministic timing for DeferrableTestCase