--- a/unittesting/core/__init__.py
+++ b/unittesting/core/__init__.py
@@ -1,6 +1,12 @@
-from .st3.runner import DeferringTextTestRunner
+from .st3.runner import DeferringTextTestRunner, AWAIT_WORKER
 from .st3.case import DeferrableTestCase
 from .st3.suite import DeferrableTestSuite
 from .loader import UnitTestingLoader as TestLoader
 
-__all__ = ["TestLoader", "DeferringTextTestRunner", "DeferrableTestCase", "DeferrableTestSuite"]
+__all__ = [
+    "TestLoader",
+    "DeferringTextTestRunner",
+    "DeferrableTestCase",
+    "DeferrableTestSuite",
+    "AWAIT_WORKER"
+]