--- a/unittesting/core/st3/runner.py
+++ b/unittesting/core/st3/runner.py
@@ -12,10 +12,6 @@ def defer(delay, callback, *args, **kwargs):
 class DeferringTextTestRunner(TextTestRunner):
     """This test runner runs tests in deferred slices."""
 
-    # How can this be user configurable?
-    # controlled_timing = False
-    controlled_timing = True
-
     def run(self, test):
         """Run the given test case or test suite."""
         self.finished = False
@@ -24,12 +20,8 @@ class DeferringTextTestRunner(TextTestRunner):
         result.failfast = self.failfast
         result.buffer = self.buffer
         startTime = time.time()
-        original_set_timeout_async = sublime.set_timeout_async
 
         def _start_testing():
-            if self.controlled_timing:
-                sublime.set_timeout_async = sublime.set_timeout
-
             with warnings.catch_warnings():
                 if self.warnings:
                     # if self.warnings is set, use it to filter all the warnings
@@ -95,9 +87,6 @@ class DeferringTextTestRunner(TextTestRunner):
             raise e
 
         def _stop_testing():
-            if self.controlled_timing:
-                sublime.set_timeout_async = original_set_timeout_async
-
             with warnings.catch_warnings():
                 stopTestRun = getattr(result, 'stopTestRun', None)
                 if stopTestRun is not None: