@randy3k 2018-11-28
Um, didn’t read your post carefully at the first glance. Coverage indeed doesn’t work with the sublime worker thread, I don’t have any solution except using `threading` library