@kaste 2018-11-29
I pushed this PR. There is a long explainer in https://github.com/SublimeText/UnitTesting/pull/146/commits/63b08cce81e04a6d11920c99af6e40e3c0613802

Goals:

- Overall faster tests
- Predictable async tests. If the developer knows their code 'hops' twice, we need to `yield` twice 
- Correct coverage for deferred work


I compare this PR and current master. The unit under test are (a) basically GitSavvy/#1056 and (b) the SublimeLinter test suite. 


```
Note: controlled_timing is the wrong name. controlled_timing=True means
everything runs on the main thread, and set_timeout_async is patched.
controlled_timing=False means everything runs on the worker thread, t.i.
the framework 'unittesting' runs on the worker thread, nothing is patched.


SublimeLinter
-------------

current master  test_user_args  93 tests   6,3s
                package        194 tests  14s      40% coverage       (A)

this PR controlled_timing=True
                test_user_args  93 tests   2s      
                package        194 tests  11s      40% coverage       (B)

this PR controlled_timing=False
                test_user_args  93 tests   4s       
                package        194 tests  16s      29% coverage       (C)

Note: test suite is focused on backend stuff, not really view related


GitSavvy/#1056
--------------

current master  test_graph_view  9 tests  7s      50%/52% coverage    (D)

this PR controlled_timing=True
                test_graph_view  9 tests  2.4s    69%/95% coverage    (E)

this PR controlled_timing=False
                test_graph_view  9 tests  4.3s    47%/52% coverage    (F)

Note: the unit under test are in two files so I show x%/y% coverage.
The correct coverage is 69/95


GitSavvy/#1056 optimized*
-------------------------

this PR controlled_timing=True
                test_graph_view  9 tests  1.2s    69%/95% coverage    (G)
          
this PR controlled_timing=False
                test_graph_view  9 tests  2.3s    47%/52% coverage    (H)

* E.g. instead of `yield <condition>` I just `yield` twice. 
```

Summary: this PR runs against the GitSavvy and the SublimeLinter test suite
unmodified. So it is not breaking wildly. `controlled_timing=True` gives
the correct coverage, at the same time giving best performance.


```
Notable: 

(D) slow and wrong coverage
(C) (F) (H)  controlled_timing=False  all have wrong coverage; they 
intentionally run only on the worker
(C) is slower than (A)