@kaste 2018-11-29
I want to make another comment. Between (E) and (G). The tests in the GitSavvy PR (E) mostly use `yield <condition>` or `yield <delay>`. (G) uses plain `yield` (None) and `yield <delay>`. Generally with async code, we have this three bad options. 

With just `yield` we depend on an implementation detail in that we know how often we 'hop'. It is not generally advisable to rely on that but it is good to have. 

`yield <condition>` depends on the actual side-effect to be observable. If the view changes usually this condition breaks. E.g. we depend on a *concrete* rendering but we only want to know that we rendered at all. 

With `yield <delay>` we depend on the CPU. It's basically just number guessing and is known to be flaky.