@Thom1729 2018-11-30
I can probably take a look this weekend, or possibly tomorrow.

I am concerned about patching the `sublime` module. Surely there must be another way to accomplish the same end.