@Thom1729 2018-12-02
So is the purpose of patching `sublime.set_timeout` to change the behavior of the tested code (to run everything on the main thread)? I'm not sure that this is a good idea for a unit test. The test should run the code in the way that it will be run in practice or as closely as possible.

Notwithstanding the above, patching the `sublime` module is probably a bad idea. That module is shared by every package, and patching it is likely to cause problems that are very hard to detect, troubleshoot, and fix.

I recognize the coverage issue. I would suggest that breaking the code into smaller pieces might make it easier to unit-test those pieces without worrying about async issues. Alternatively, you could use dependency injection — pass a `defer` method into your code instead of calling `sublime.set_timeout_async`.