@kaste 2018-12-03
We should explore the design space we have here. Your 'concerns' are very broad and general opinions.

We have at least two problems/issues to solve: 

a) The tests are too slow. (E.g. the GitSavvy tests: 9 tests in 7 seconds doesn't scale, and you basically can't practice TDD in a meaningful way.)

b) The coverage is completely off. (The workaround would be to write tests for the coverage not around the features you actually want to ensure and protect.)

As a plus, there is a third issue. This is a reasonable issue bc we already have a DefferableTestCase with `yield` which promises to make testing 'async' code better/easier/feasable. So the issue is:

c) Plain `yield` (aka `yield None`) has no meaningful behavior. (Currently, it directly translates to `yield 10` [milliseconds], but 10ms has no meaningful semantics for a developer.) We should be able to tick the clock in a controlled way using `yield`.

  
For this PR:

- Is there a minimal commit/changeset we can ship for all users which has a positive impact without breaking anything.

- Is there a way to apply some of these changes a la carte? Does it make sense to turn these on/off for the scope of a class or a single test?

- Can we still have a meaningful `yield 'AWAIT_WORKER_THREAD'`?

- How can we deploy something like this so that *travis* can run these changes? E.g. can I teach *travis* to run with master and a next version of unittesting so we can create some confidence over time?