@Thom1729 2018-12-03
As you say, it's tricky to try to solve all of those problems at once. So let's look at them one by one:

> a) The tests are too slow. (E.g. the GitSavvy tests: 9 tests in 7 seconds doesn't scale, and you basically can't practice TDD in a meaningful way.)

Why are the GitSavvy tests slow? Would a less drastic solution make them faster? (For instance, using set_timeout in the runner instead of set_timeout_async, but not patching `sublime` to do that globally.)

> b) The coverage is completely off. (The workaround would be to write tests for the coverage not around the features you actually want to ensure and protect.)

This is a general problem with `coverage`. It could be partly ameliorated by using `set_timeout` in the runner, but that doesn't help if the GitSavvy code uses `set_timeout_async` directly. The approach you propose would cause GitSavvy to use `set_timeout` instead.

However, you could achieve the same result by having an explicit switch inside GitSavvy that chooses which function to use. This would be much safer than patching the shared `sublime` object. There is a risk to having your code behave differently during testing, but that risk is no less if it behaves differently due to a patched `sublime` object.

> c) Plain `yield` (aka `yield None`) has no meaningful behavior. (Currently, it directly translates to `yield 10` [milliseconds], but 10ms has no meaningful semantics for a developer.) We should be able to tick the clock in a controlled way using `yield`.

This sounds less like a problem and more like a place that a feature could be added. I'm not sure that this is really related to the other questions.

---

I don't mean to be dismissive, but I don't think that we can "create some confidence over time" for monkeypatching the `sublime` object. That might be an absolute last-resort measure if it were impossible to implement core functionality without it. Here, though, it seems that there are workarounds.