@kaste 2018-12-04
Yes, of course, the performance gains come from changing the timings in the `runner`. However, if you *just* lower these values you get sync problems and your tests get flaky. It is basically a feature that when you run your tests single threaded, you can actually push these values because you have deterministic timing. The tasks run ordered and in a controlled way. 

Changing the unit tests in GitSavyy or SublimeLinter has only some limited effect. I used in the example above SublimeLinter's `test_user_args.dsl.py` https://github.com/SublimeLinter/SublimeLinter/blob/master/tests/test_user_args_dsl.py which never yields and actually is 'pure' in that it doesn't interact with the Sublime API at all. As I said above, I can bring down this test from 6s to 2s for 93 tests. 

For comparison since you mentioned sublime_lib. 

```
sublime_lib master
------------------

current master   171 tests        35s
this PR controlled_timing=True     9s
this PR controlled_timing=False   55s

Note: 5 failing tests regardless how I run these tests. 2 look like bugs, 3 of them are async issues. 

this PR controlled_timing=True  & fixed and optimized tests 
                                  7,5s

```

sublime_lib maybe has a handful `yield`s so there aren't just like in SublimeLinter much options to make the unit tests itself faster. As it stands, the framework itself is slow.